backend/requirements.txt
pymysql
python-dotenv
fastapi
gtts
openai
uvicorn
gunicorn
aiomysql
python-multipart